+++ core/src/index/knowhere/knowhere/index/vector_index/FaissBaseIndex.cpp
@@ -70,7 +70,8 @@ FaissBaseIndex::SealImpl() {
     faiss::Index* index = index_.get();
     auto idx = dynamic_cast<faiss::IndexIVF*>(index);
     if (idx != nullptr) {
-        KNOWHERE_LOG_DEBUG << "Test before to_readonly:" << " IVF READONLY " << std::boolalpha << idx->is_readonly();
+        KNOWHERE_LOG_DEBUG << "Test before to_readonly:"
+                           << " IVF READONLY " << std::boolalpha << idx->is_readonly();
         idx->to_readonly();
     }
 #endif