+++ CHANGELOG.md
@@ -5,9 +5,11 @@ Please mark all change in change log and use the issue from GitHub
 # Milvus 0.7.0 (TBD)
 
 ## Bug
+- \#744 - Don't return partition table for show_tables
 
 ## Feature
 - \#343 - Add Opentracing
+- \#665 - Support get/set config via CLI
 
 ## Improvement
 - \#738 - Use openblas / lapack from apt install
@@ -105,6 +107,7 @@ Please mark all change in change log and use the issue from GitHub
 - \#584 - Intergrate internal FAISS
 - \#611 - Remove MILVUS_CPU_VERSION
 - \#634 - FAISS GPU version is compiled with O0
+- \#737 - Refactor server module to separate Grpc from server handler and scheduler
 
 ## Task